notes-to-quiz/backend/requirements.txt
flask
flask-cors
pytesseract
pillow
PyPDF2
opencv-python-headless
numpy
easyocr
torch















































































